+++ packages/citration-core-swift/Sources/CitrationCore/Remote/Models/AuthResponse.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-struct AuthResponse: Codable {
-    let accessToken: String
-    let refreshToken: String
-    let expiresAt: String
-    let user: User
-}